+++ Bank/Models/Transaction.cs
@@ -10,7 +10,7 @@ namespace BankSimulation.Models
 	
 	public class Transaction
 	{
-		public static int decimalNumberRound = 3;
+		public static Byte decimalNumberRound = 3;
 		public Transaction() {
 			this.timestamp = DateTime.Now;
 		}
